Fix issue 80: Weather packet decoding.
Weather packets were being decoded incorrectly.  Over the air packets
for weather are in wind miles per hour.  Rain is in hundreths of an
inch according to the official spec.

This patch changes the weather packet output to include the wind_speed
and wind_direction as part of the weather dict in the packet instead of
course and speed.  According to the APRS Spect CSE/SPD for weather
packets, which are denoted with the _ character are supposed to be used
for wind speed and wind direction.

"The on-air APRS WX protocols, however, still will be in MPH and F."

"RAIN VALUES:  Rain is counted in increments of 0.1 or 0.01 inch or 1mm.
but reports all values in 0.01 inches over the air."

References:
http://www.aprs.org/APRS-docs/WX.TXT